==> codes/1a - Predict Aspect categories using LinearSVM + Bag of Words.ipynb <==
# Aspect category classification 

import pandas as pd
import numpy as np
pd.set_option('display.max_colwidth' , -1)

train_grp_df = pd.read_csv('../data/resturant_train_stratified_grouped.csv')
test_grp_df  = pd.read_csv('../data/resturant_test_stratified_grouped.csv')
train_grp_df.shape , test_grp_df.shape

train_grp_df.head(2)

First step of Aspect Based Sentiment Analysis is to extract the hidden aspect categories in given text/review. We are using a supurvised method for this.  We will learn a LinearSVM on tf-idf vectors for this multi label classification task.

import ast 
train_grp_df['aspects'] = train_grp_df['aspects'].apply(lambda x: ast.literal_eval(x))
train_grp_df['polarities'] = train_grp_df['polarities'].apply(lambda x: ast.literal_eval(x))

test_grp_df['aspects'] = test_grp_df['aspects'].apply(lambda x: ast.literal_eval(x))
test_grp_df['polarities'] = test_grp_df['polarities'].apply(lambda x: ast.literal_eval(x))

train_grp_df.head(2)

from sklearn.feature_extraction.text import CountVectorizer
from nltk.corpus import stopwords
from nltk.stem.wordnet import WordNetLemmatizer
import codecs

def parse_sentence(line):
    lmtzr = WordNetLemmatizer()
    stop = stopwords.words('english')
    try:
        text_token = CountVectorizer().build_tokenizer()(line.lower())
        text_rmstop = [i for i in text_token if i not in stop]
        text_stem = [lmtzr.lemmatize(w) for w in text_rmstop]
        return text_stem
    except:
        return []


train_grp_df['pro_text'] = train_grp_df['text'].apply(lambda x: parse_sentence(x))
test_grp_df['pro_text'] = test_grp_df['text'].apply(lambda x: parse_sentence(x))

train_grp_df['clean_text'] = train_grp_df['pro_text'].apply(lambda x: ' '.join(x))
test_grp_df['clean_text'] = test_grp_df['pro_text'].apply(lambda x: ' '.join(x))

Preprocess text by tokenizing and lemmatizing then create tf-idf vectors on processed text. Then we create a multi label classifier using Linear SVM to predict aspects.

train_grp_df.head(2)

from sklearn.feature_extraction.text import CountVectorizer , TfidfVectorizer ,TfidfTransformer
from sklearn.pipeline import Pipeline
from skmultilearn.problem_transform import LabelPowerset
from sklearn.linear_model import SGDClassifier

# Convert the multi-labels into arrays
from sklearn.preprocessing import MultiLabelBinarizer
import pickle
mlb = MultiLabelBinarizer()
y_train = mlb.fit_transform(train_grp_df.aspects)
y_test = mlb.fit_transform(test_grp_df.aspects)


y_train[0:5]

text_clf_svm = Pipeline([('vect', CountVectorizer()),
                         ('tfidf', TfidfTransformer()),
                         ('clf-svm', LabelPowerset(
                             SGDClassifier(loss='log', penalty='l2', l1_ratio = 0.4,
                                           alpha=1e-4, max_iter=40, random_state=42 )))])
_ = text_clf_svm.fit(train_grp_df['clean_text'] , y_train)
predicted_svm = text_clf_svm.predict(test_grp_df['clean_text'])
print('Test accuracy' , np.mean(predicted_svm == y_test))

"""
Function to tweak threshold and determine aspect prediction using that threshold
"""

def make_pred(row , thresh):
    cols = mlb.classes_.tolist()
    res_asp = []
    for ele in cols:
        if row[ele]>thresh:
            res_asp.append(ele)

    if len(res_asp)==0:
        best_sr = cols[np.argmax(row[cols].tolist())]
        res_asp.append(best_sr)
        return res_asp
    return res_asp
        

test_prob_df = pd.DataFrame(text_clf_svm.predict_proba(test_grp_df['clean_text']).toarray() , columns = mlb.classes_ , index = test_grp_df.index)
test_grp_df= pd.merge(test_grp_df, test_prob_df , left_index = True , right_index=True)

test_grp_df['aspects_pred'] =test_grp_df.apply(lambda x: make_pred(x ,0.5) , axis=1)

test_grp_df[['text' , 'aspects' , 'aspects_pred']].head()

from sklearn.metrics import f1_score , confusion_matrix , accuracy_score , precision_score , recall_score , roc_auc_score
print("AUC ROC"  , roc_auc_score(y_test ,text_clf_svm.predict_proba(test_grp_df['clean_text']).toarray() ))#
print("F1 score",f1_score(y_test ,mlb.transform(test_grp_df.aspects_pred) , average='macro' ))
print('Overall accuracy' , np.mean(predicted_svm == y_test))

print('Accuracy for different aspect categories ')
print("ambience" ,   accuracy_score( y_test[: , 0] ,mlb.transform(test_grp_df.aspects_pred)[:,0] ))
print("anecdotes/miscellaneous" ,   accuracy_score( y_test[: , 1] ,mlb.transform(test_grp_df.aspects_pred)[:,1] ))
print("food" ,   accuracy_score( y_test[: , 2] ,mlb.transform(test_grp_df.aspects_pred)[:,2] ))
print("price" ,   accuracy_score( y_test[: , 3] ,mlb.transform(test_grp_df.aspects_pred)[:,3] ))
print("service" ,   accuracy_score( y_test[: , 4] ,mlb.transform(test_grp_df.aspects_pred)[:,4] ))

### Look for the accuracy of conflicting statements (changing sentiment in single review for different aspects)


test_grp_df2 = test_grp_df[test_grp_df['ind']==1]
y_test_conf = mlb.fit_transform(test_grp_df2.aspects)

print("F1 score",f1_score(y_test_conf ,mlb.transform(test_grp_df2.aspects_pred) , average='macro' ))
print('Overall accuracy' , np.mean(text_clf_svm.predict(test_grp_df2['clean_text']) == y_test_conf))

print('Accuracy for different aspect categories ')
print("ambience" ,   accuracy_score( y_test_conf[: , 0] ,mlb.transform(test_grp_df2.aspects_pred)[:,0] ))
print("anecdotes/miscellaneous" ,   accuracy_score( y_test_conf[: , 1] ,mlb.transform(test_grp_df2.aspects_pred)[:,1] ))
print("food" ,   accuracy_score( y_test_conf[: , 2] ,mlb.transform(test_grp_df2.aspects_pred)[:,2] ))
print("price" ,   accuracy_score( y_test_conf[: , 3] ,mlb.transform(test_grp_df2.aspects_pred)[:,3] ))
print("service" ,   accuracy_score( y_test_conf[: , 4] ,mlb.transform(test_grp_df2.aspects_pred)[:,4] ))

The overall accuracy for test data is around 86% and F1 score is 66% . We will try some LSTM + Word embedding based techniques to improve the performace of the multi label classification problem.
