==> codes/1b- Sentence segmentation using constituency parsing.ipynb <==
import pandas as pd
import numpy as np
pd.set_option('display.max_colwidth', -1)

import os
import pandas as pd
import numpy as np
import pickle
from collections import Counter, defaultdict
import re
import spacy
from spacy import displacy
#import gensim
#import nltk


test_aspect_pred = pd.read_csv('../output/BiLSTM_aspect_prediction_test_data.csv')

import ast 
test_aspect_pred['aspects_pred'] = test_aspect_pred['aspects_pred'].apply(lambda x: ast.literal_eval(x))
test_aspect_pred['aspects'] = test_aspect_pred['aspects'].apply(lambda x: ast.literal_eval(x))

test_aspect_pred['ct_aspects'] = test_aspect_pred['aspects_pred'].apply(lambda x: len(x))
test_aspect_pred['ct_aspects'].value_counts()

test_aspect_pred[test_aspect_pred['ct_aspects']==0].tail()

##### We will try to break the review text using constituency parsing . If the reviews contain constrating conjuction and they have not been split by constituency parsing method , we will try to split them using constrasting conjuction.

import time
import benepar

def traverse_tree(sentence):
    res = []
    parser = benepar.Parser("benepar_en2")
    tree = parser.parse(sentence)
    for subtree in tree:
        if type(subtree) == nltk.tree.Tree:
            type_list = ['S' ] 
            if  subtree.label() in type_list :            
                res.append(subtree.leaves())
    if len(res)==0:
        res = [sentence.split(' ')]
    for ele in res:
        if isinstance(ele , list)  and len(ele)==1:
            res.remove(ele)
    return res

import time
st = time.time()
test_aspect_pred['splits'] = test_aspect_pred.apply(lambda x: traverse_tree(x['text']) if x['ct_aspects'] >1 else [x['text'].split(' ')] , axis = 1 )
en = time.time()
print('time taken in minutes',(en -st)/60)

If reading the saved csv file, need to convert splits column again using ast literal eval 

#test_aspect_pred['splits']= test_aspect_pred['splits'].apply(lambda x: ast.literal_eval(x))
test_aspect_pred['length_split'] = test_aspect_pred['splits'].apply(lambda x: len(x))
test_aspect_pred.length_split.value_counts()

test_aspect_pred.tail()

import ast
#test_aspect_pred.to_csv('../output/test_data_sentence_segmentation.csv' , index = False)
test_aspect_pred = pd.read_csv('../output/test_data_sentence_segmentation.csv')
for col in ['aspects_pred' , 'aspects' , 'splits']:
    test_aspect_pred[col]  = test_aspect_pred[col].apply(lambda x: ast.literal_eval(x))


test_aspect_pred.tail(2)

test_aspect_pred['length_splits'] = test_aspect_pred['splits'].apply(lambda x: len(x))
test_aspect_pred.length_splits.value_counts()

single_aspect_df = test_aspect_pred[(test_aspect_pred['length_splits']==1)] # & (test_aspect_pred['ct_aspects']<=1)
multi_aspect_df = test_aspect_pred[test_aspect_pred['length_splits']>1]

multi_aspect_df.drop(columns=['aspects' , 'ct_aspects', 'length_splits' ] , inplace=True)

text_aspectwise= []
for index, row in multi_aspect_df.iterrows():                   
    total_aspects = row['aspects_pred'].copy()
    total_sent = row['splits'].copy()    
    for sent in total_sent:
        text_aspectwise.append({  'id': index , 'sp_text': ' '.join(sent) , 'aspects_pred': total_aspects })

df_test_aspectwise = pd.DataFrame(text_aspectwise)

df_test_aspectwise.shape

import pickle 
new_vocab = pickle.load(open('../../../data/manually_tweaked_dict.pickle' , 'rb'))
new_vocab['price'].append('pricy')
new_vocab['price'].append('dollar')

len(new_vocab['price'])

new_vocab.keys()




        for aspect , polarity in zip(total_aspects ,total_polarities):
            match_words = []
            for sen_list in row['conj_splits']:
                if aspect in sen_list:
                    match_words.append(999)  ## Exact word present in the sentence
                else:
                    words = sum(word.lower() in new_vocab[aspect] for word in sen_list)
                    match_words.append(words)

df_test_aspectwise

test_preds = []
text_data = []
true_label = []

epoch_loss = 0
epoch_acc = 0 
ct = 0
model.eval()
with torch.no_grad():
    for x ,y  in test_loader:
        text, text_lengths = x
        text_data.append(text.data.cpu().numpy())
        predictions = model(text, text_lengths)#.squeeze(1)
        rounded_preds = torch.round(torch.sigmoid(predictions))  #torch.round
        preds = rounded_preds.data.cpu().numpy()
        test_preds.append(preds)
        true_label.append(y.data.cpu().numpy())


import pickle 
new_vocab = pickle.load(open('../../data/manually_tweaked_dict.pickle' , 'rb'))
new_vocab['price'].append('pricy')
new_vocab['price'].append('dollar')

"""
some of the lists  are empty/ or contains ' ' characters only. Removing them using following function
"""
def remove_null(lists):
    for lst in lists:
        if isinstance(lst , list) & ( lst ==[''] or lst==[' '] ) :
            lists.remove(lst)
    return lists 

test_grp_df['conj_splits'] = test_grp_df['conj_splits'].apply(lambda x: remove_null(x))
test_grp_df['length_split'] = test_grp_df['conj_splits'].apply(lambda x: len(x))
test_grp_df.length_split.value_counts()

test_grp_df.reset_index(inplace=True , drop = True)
test_grp_df['aspects'] = test_grp_df['aspects'].apply(lambda x: [i for i in x.split(' ')])
test_grp_df['polarities'] = test_grp_df['polarities'].apply(lambda x: [i for i in x.split(' ')])

Once we have sentence fragments, we need to map each sentence fragments to corresponding aspect categories. Created a dictionary containing key words corresponding to each aspect categories. And implementing following steps: 

1) if only one fragment in splits , assign complete sentence to each of the aspect categories. 

2) For each aspect category, look at each fragment  and count # of keywords present. Assign fragment with higher matching keywords to given aspect.

3) if we have zero matching keyword in all the sentence fragments, assign the complete sentenence to that aspect category.

text_aspectwise= []
for index , row in test_grp_df.iterrows():

    if row['length_split']==1:
        for aspect , polarity in zip(row['aspects'] , row['polarities']):
            text_aspectwise.append({'id':index ,'text':row['text'].lower(), 'aspect': aspect,'polarity': polarity})
            
    else:                   
        total_aspects = row['aspects'].copy();   total_polarities = row['polarities'].copy()
        for aspect , polarity in zip(total_aspects , total_polarities):
            if aspect == 'anecdotes/miscellaneous':
                text_aspectwise.append({'id':index ,'text':row['text'].lower(), 'aspect': aspect, 'polarity': polarity})
                total_aspects.remove(aspect); total_polarities.remove(polarity)
                break

        total_sent = row['conj_splits'].copy()
        for aspect , polarity in zip(total_aspects ,total_polarities):
            match_words = []
            for sen_list in row['conj_splits']:
                if aspect in sen_list:
                    match_words.append(999)  ## Exact word present in the sentence
                else:
                    words = sum(word.lower() in new_vocab[aspect] for word in sen_list)
                    match_words.append(words)
            if np.max(match_words)!=0:
                best_sent = row['conj_splits'][np.argmax(match_words)]
                text_aspectwise.append({'id':index,'text':' '.join(best_sent),   'aspect': aspect,'polarity': polarity})
            else:
                
                text_aspectwise.append({'id':index,'text':row['text'],   'aspect': aspect,'polarity': polarity})
                


df_aspectwise_test =    pd.DataFrame.from_records(text_aspectwise)
df_aspectwise_test.to_csv('../../data/test_data_aspectXtext_fragment.csv')
